Body radius is now a float and is displayed in correct units in overview details

+++ Source/MiniSolarSystem/Private/GasGiant.cpp
@@ -27,7 +27,7 @@ void AGasGiant::BeginPlay()
 	ColorSettings->NewVoronoiForStorms();
 }
 
-void AGasGiant::SetRadius(int NewRadius)
+void AGasGiant::SetRadius(float NewRadius)
 {
 	Radius = NewRadius;
 	this->SetActorScale3D(FVector(Radius));
